feat: decomposition files ngrx
