+++ src/app/core/core.module.ts
@@ -6,9 +6,9 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { StoreModule } from '@ngrx/store';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { EffectsModule } from '@ngrx/effects';
-import { reducers, metaReducers } from './reducers';
+import { reducers, metaReducers } from '../redux/reducers';
 import { environment } from '../../environments/environment';
-import { CoreEffects } from './core.effects';
+import { CoreEffects } from '../redux/effects/core.effects';
 import { CoreComponent } from './core.component';
 import { HeaderComponent } from './header/header.component';
 import { FooterComponent } from './footer/footer.component';
