+++ src/app/redux/keys.ts
@@ -1,0 +1,3 @@
+export const FILTER_KEY = 'filter';
+export const MADE_ITEM_KEY = 'madeItem';
+export const HEADER_KEY = 'header';
