+++ src/app/redux/reducers/index.ts
@@ -3,19 +3,20 @@ import {
   MetaReducer,
 } from '@ngrx/store';
 import { environment } from '../../../environments/environment';
-import { filterReducer, FilterState, FILTER_KEY } from './filters';
+import { filterReducer, FilterState } from './filters';
 import { headerReducer, HeaderState } from './header';
-import { MADE_ITEM_KEY, madeItemReducer, MadeItemState } from './madeItem';
+import { madeItemReducer, MadeItemState } from './madeItem';
+import { MADE_ITEM_KEY, HEADER_KEY, FILTER_KEY } from '../keys';
 
 export interface State {
   [FILTER_KEY]: FilterState;
-  header: HeaderState;
+  [HEADER_KEY]: HeaderState;
   [MADE_ITEM_KEY]: MadeItemState;
 }
 
 export const reducers: ActionReducerMap<State> = {
   [FILTER_KEY]: filterReducer,
-  header: headerReducer,
+  [HEADER_KEY]: headerReducer,
   [MADE_ITEM_KEY]: madeItemReducer,
 };
