+++ src/app/youtube/components/filtering/filtering.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { isShowFilterSelector } from '../../../core/reducers/filters';
+import { isShowFilterSelector } from '../../../redux/selectors/filters';
 
 @Component({
   selector: 'app-filtering',
